+++ homeassistant/components/androidtv/__init__.py
@@ -1,0 +1,6 @@
+"""
+Support for functionality to interact with Android TV and Fire TV devices.
+
+For more details about this platform, please refer to the documentation at
+https://home-assistant.io/components/media_player.androidtv/
+"""
