+++ homeassistant/components/androidtv/services.yaml
@@ -1,0 +1,11 @@
+# Describes the format for available Android TV and Fire TV services
+
+adb_command:
+  description: Send an ADB command to an Android TV / Fire TV device.
+  fields:
+    entity_id:
+      description: Name(s) of Android TV / Fire TV entities.
+      example: 'media_player.android_tv_living_room'
+    command:
+      description: Either a key command or an ADB shell command.
+      example: 'HOME'
