--- homeassistant/components/firetv/__init__.py
@@ -1,6 +1,0 @@
-"""
-Support for functionality to interact with FireTV devices.
-
-For more details about this platform, please refer to the documentation at
-https://home-assistant.io/components/media_player.firetv/
-"""
