--- homeassistant/components/firetv/services.yaml
@@ -1,11 +1,0 @@
-# Describes the format for available Fire TV services
-
-adb_command:
-  description: Send an ADB command to a Fire TV device.
-  fields:
-    entity_id:
-      description: Name(s) of Fire TV entities.
-      example: 'media_player.fire_tv_living_room'
-    command:
-      description: Either a key command or an ADB shell command.
-      example: 'HOME'
